backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
aiosqlite>=0.19.0

# AI providers
openai>=1.10.0
httpx>=0.26.0

# GitHub
pygithub>=2.1.1
gitpython>=3.1.42

# Configuration
pyyaml>=6.0.1
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Logging and monitoring
structlog>=24.1.0
python-json-logger>=2.0.7

# Utilities
aiofiles>=23.2.1
tenacity>=8.2.3
